Fix PGM options scanning
If an unknown option is encountered in the scan of the PGM options in a
packet then the bp pointer is set too far ahead. The scanning of the
following options will fail.

tcpdump will only display the first few options without this fix.

This is due to the bp pointer being incremented twice to process
the options type and length even if the option is unknown.

If this is not taken into account then the bp pointer will point to the
contents of the next pgm option and not to the type and length as
expected and tcpdump will print an error. For example:

00:02:55.992148 IP (tos 0x0, ttl 16, id 58109, offset 0, flags [DF],
proto UDP (17), length 122)
    192.168.193.237.63846 > 239.90.0.26.19026: 26198 > 19026: PGM,
length 0 0xc0a85322cff9 SPM seq 3835 trail 1 lead 545 nla
192.168.193.237 OPTS LEN 58 OPT_21 [30] [Bad option, length 0 < 4] [94]

Signed-off-by: Christoph Lameter <cl@linux.com>