PGM: Verify ODATA sequence numbers
The primary purpose of the PGM protocol is the delivery of packets in
sequence. PGM allows the recovery of lost messages and the elimination
of duplicate messages via sequence numbers for the individual packets.

It is therefore useful to flag ODATA packets that are out of order.
For that purpose tcpdump needs to remember the current
sequence number for each GSI (Global Sender ID). Then we can flag
if there are sequence errors. These are usually the focus of
investigations when troubleshooting PGM issues on the network.

Output sample where one packet is missing between SQN 42839 und 42837:

 192.168.193.242.52629 > 239.91.0.216.39217: 18554 > 39217: PGM, length
231 0xc0a8c1f2c9ef [Missing ODATA! lead 42837]ODATA trail 1 seq 42839
OPTS LEN 57 OPT_21 [30]  OPT_29 [8]  OPT_A2 [15]  [312]
00:10:51.731519 IP (tos 0x0, ttl 16, id 35576, offset 0, flags [+],
proto UDP (17), length 1500)

Signed-off-by: Christoph Lameter <cl@linux.com>